add 'int' as sql alias for 'bigint' patch by jbellis
git-svn-id: https://svn.apache.org/repos/asf/cassandra/trunk@1090798 13f79535-47bb-0310-9956-ffa450edef68

+++ test/system/test_cql.py
@@ -68,7 +68,7 @@ def load_sample(dbconn):
             WITH comparator = ascii AND default_validation = uuid;
     """)
     dbconn.execute("""
-        CREATE COLUMNFAMILY IndexedA (KEY text PRIMARY KEY, birthdate bigint)
+        CREATE COLUMNFAMILY IndexedA (KEY text PRIMARY KEY, birthdate int)
             WITH comparator = ascii AND default_validation = ascii;
     """)
     dbconn.execute("CREATE INDEX ON IndexedA (birthdate)")
